refactor(history): extract HistoryPanel + share time/wiki utils (DRY)
page.tsx was a 2700-line dumping ground holding the history panel,
its row component, source-trailer parsing, three near-duplicate time
formatters, and inline history types/fetch. Pulled apart into proper
shared architecture:

- src/lib/time.ts (new): single relativeTime(iso, 'short'|'long')
  util replacing the duplicated formatRelative + formatRelativeLong.
  Plus absoluteTime() for tooltips. One source of truth for
  'N minutes ago' across the app.
- src/lib/wiki.ts: CommitInfo + FileHistoryResponse types,
  fetchFileHistory(), and parseCommitSource() (was the inline
  parseSourceMeta) now live with the rest of the wiki API client.
- src/components/wiki/HistoryPanel.tsx (new): HistoryPanel + private
  ActivityRow + Avatar, importing the shared utils. Self-contained,
  testable in isolation.
- page.tsx: drops the moved code (-360 lines), imports HistoryPanel
  + relativeTime/absoluteTime + fetchFileHistory, swaps the document
  list and active-agents-bar formatters onto the shared relativeTime.

Net: page.tsx −345 lines; history surface is now reusable, DRY, and
each file has one responsibility. No behavior change — render
verified identical (author title, verbose relative time, no commit
message in the sidebar).
